v. 2.0.3

+++ predictPCRartifacts.R
@@ -43,7 +43,6 @@ sites <- bind_rows(lapply(split(sites, sites$refGenome), function(x){
        x$n <- ntile(1:nrow(x), opt$predictPCRartifacts_CPUs)
        
       r <- bind_rows(parLapply(cluster, split(x, x$n), function(a){
-      #r <- bind_rows(lapply(split(x, x$n), function(a){   
               library(dplyr)
               library(Biostrings)
               library(rtracklayer)
@@ -68,8 +67,6 @@ sites <- bind_rows(lapply(split(sites, sites$refGenome), function(x){
                 leaderSeq <- subset(x, posid == x2)$repLeaderSeq
                 i <- ifelse(nchar(leaderSeq) < opt$predictPCRartifacts_adjacentSeqLength, nchar(leaderSeq),opt$predictPCRartifacts_adjacentSeqLength)
                 t <- substr(leaderSeq, nchar(leaderSeq)-i+1, nchar(leaderSeq))
-             
-                #browser()
                 
                 if(strand == '-'){
                   endPos <- pos + opt$predictPCRartifacts_adjacentSeqLength
@@ -134,7 +131,6 @@ sites <- bind_rows(lapply(split(sites, sites$refGenome), function(x){
        left_join(x, r, by = 'uniqueSite') %>% relocate(PCRartifact1, .after = opt$predictPCRartifacts_addAfter)
 }))
 
-
 o <- bind_rows(lapply(split(sites, sites$refGenome), function(x){
   
   g <- import(rtracklayer::TwoBitFile(x$refGenome[1]))
@@ -180,22 +176,21 @@ o <- bind_rows(lapply(split(sites, sites$refGenome), function(x){
   left_join(x, r, by = 'uniqueSite')
 }))
 
-
 o2 <- bind_rows(lapply(split(o, o$vectorFastaFile), function(a){
   vectorFastaFile <- a$vectorFastaFile[1]
-  
+    
   invisible(file.remove(list.files(file.path(opt$outputDir, opt$predictPCRartifacts_outputDir, 'dbs'), full.names = TRUE)))
   system(paste0('makeblastdb -in ', vectorFastaFile, ' -dbtype nucl -out ', file.path(opt$outputDir, opt$predictPCRartifacts_outputDir, 'dbs', 'd')), ignore.stderr = TRUE)
   
   cluster <- makeCluster(opt$predictPCRartifacts_CPUs)
   clusterExport(cluster, c('opt'))
   
-  r <- bind_rows(parLapply(cluster, split(a, 1:nrow(a)), function(x){
+   r <- bind_rows(parLapply(cluster, split(a, 1:nrow(a)), function(x){
     library(dplyr)
     library(Biostrings)
     library(stringr)
     source(file.path(opt$softwareDir, 'lib.R'))
-    
+        
     f <- tmpFile()
     
     write(c('>seq', x$seq), file = file.path(opt$outputDir, opt$predictPCRartifacts_outputDir, 'dbs', f), append = FALSE)
@@ -205,8 +200,7 @@ o2 <- bind_rows(lapply(split(o, o$vectorFastaFile), function(a){
                   file.path(opt$outputDir, opt$predictPCRartifacts_outputDir, 'dbs', f), ' -db ',
                   file.path(opt$outputDir, opt$predictPCRartifacts_outputDir, 'dbs', 'd'),
                   ' -out ', file.path(opt$outputDir, opt$predictPCRartifacts_outputDir, 'dbs', paste0(f, '.blast'))), ignore.stdout = TRUE, ignore.stderr = TRUE)
-    
-    
+        
     if(file.info(file.path(opt$outputDir, opt$predictPCRartifacts_outputDir, 'dbs', paste0(f, '.blast')))$size > 0){
       
       b <- read.table(file.path(opt$outputDir, opt$predictPCRartifacts_outputDir, 'dbs', paste0(f, '.blast')), sep = '\t', header = FALSE)
@@ -215,9 +209,9 @@ o2 <- bind_rows(lapply(split(o, o$vectorFastaFile), function(a){
       b$matches = b$length - b$mismatch
       
       b <- dplyr::filter(b, matches >= opt$predictPCRartifacts_minReportHalfMatches) %>% dplyr::slice_max(bitscore, with_ties = TRUE)
-      
+            
       if(nrow(b) > 0){
-        
+                
         b <- dplyr::slice_max(b, matches, with_ties = FALSE)
         
         f2 <- tmpFile()
@@ -225,49 +219,65 @@ o2 <- bind_rows(lapply(split(o, o$vectorFastaFile), function(a){
         v <- v[names(v) == b$sseqid]
         writeXStringSet(v, file.path(opt$outputDir, opt$predictPCRartifacts_outputDir, 'dbs', paste0(f2, '.fasta')))
         system(paste0('makeblastdb -in ', file.path(opt$outputDir, opt$predictPCRartifacts_outputDir, 'dbs', paste0(f2, '.fasta')), ' -dbtype nucl -out ', file.path(opt$outputDir, opt$predictPCRartifacts_outputDir, 'dbs', f2)), ignore.stderr = TRUE)
-        
+                
         system(paste0('blastn -dust no -soft_masking false -evalue 50 -outfmt 1 -word_size ', opt$predictPCRartifacts_wordSize, 
                       ' -gapopen ', opt$predictPCRartifacts_gapOpeningPenalty, ' -gapextend ', opt$predictPCRartifacts_gapExtensionPenalty, ' -query ',
                       file.path(opt$outputDir, opt$predictPCRartifacts_outputDir, 'dbs', f), ' -db ',
                       file.path(opt$outputDir, opt$predictPCRartifacts_outputDir, 'dbs', f2),
                       ' -out ', file.path(opt$outputDir, opt$predictPCRartifacts_outputDir, 'dbs', paste0(f, '.blast2'))), ignore.stdout = TRUE, ignore.stderr = TRUE)
+                
+        if(file.info(file.path(opt$outputDir, opt$predictPCRartifacts_outputDir, 'dbs', paste0(f, '.blast2')))$size > 0){
+                    
+          p <- readLines(file.path(opt$outputDir, opt$predictPCRartifacts_outputDir, 'dbs', paste0(f, '.blast2')))
+          invisible(file.remove(list.files(file.path(opt$outputDir, opt$predictPCRartifacts_outputDir, 'dbs'), pattern = f2, full.names = TRUE)))
         
-        p <- readLines(file.path(opt$outputDir, opt$predictPCRartifacts_outputDir, 'dbs', paste0(f, '.blast2')))
-        invisible(file.remove(list.files(file.path(opt$outputDir, opt$predictPCRartifacts_outputDir, 'dbs'), pattern = f2, full.names = TRUE)))
+          if(length(p) == 0){
+            invisible(file.remove(list.files(file.path(opt$outputDir, opt$predictPCRartifacts_outputDir, 'dbs'), pattern = f, full.names = TRUE)))
+            return(tibble(uniqueSite = x$uniqueSite, PCRartifact2 = NA))
+          }
+                    
+          q <- p[grepl('Query_1', p)]
+          
+          if(length(q) == 0){
+            invisible(file.remove(list.files(file.path(opt$outputDir, opt$predictPCRartifacts_outputDir, 'dbs'), pattern = f, full.names = TRUE)))
+            return(tibble(uniqueSite = x$uniqueSite, PCRartifact2 = NA))
+          }
+                  
+          dashes <- stringr::str_count(str_extract(q, '[ATCGN\\-]+'), '-')
+                    
+          i <- stringr::str_locate(q, '[ATCGN\\-]+')
+          q <- as.integer(unlist(stringr::str_extract_all(sub('Query_1\\s+', '', q), '\\d+')))
         
-        q <- p[grepl('Query_1', p)]
+          p <- p[grepl(paste0('\\.\\s+', b$send, '$'), p)][1]
+          p <- substr(p, i[1], i[2])
+          p <- gsub('\\.', '*', p)
+          p <- gsub('\\s', '.', p)
+          p <- unlist(strsplit(p, ''))
+          names(p) <- q[1]:(q[2]+dashes)
         
-        dashes <- stringr::str_count(str_extract(q, '[ATCGN\\-]+'), '-')
-        i <- stringr::str_locate(q, '[ATCGN\\-]+')
-        q <- as.integer(unlist(stringr::str_extract_all(sub('Query_1\\s+', '', q), '\\d+')))
+          p <- paste0(unlist(lapply(1:(opt$predictPCRartifacts_adjacentSeqLength+dashes), function(n){
+            n <- as.character(n)
+            if(n %in% names(p)) return(p[names(p) == n])
+            return('.')
+          })), collapse = '')
         
-        p <- p[grepl(paste0('\\.\\s+', b$send, '$'), p)][1]
-        p <- substr(p, i[1], i[2])
-        p <- gsub('\\.', '*', p)
-        p <- gsub('\\s', '.', p)
-        p <- unlist(strsplit(p, ''))
-        names(p) <- q[1]:(q[2]+dashes)
+          # Look at the right half of the alignment and report it if there are >= opt$predictPCRartifacts_minReportHalfMatches matches.
+          t <- substr(p, ceiling(opt$predictPCRartifacts_adjacentSeqLength / 2)+1, nchar(p))
+          
+          if((stringr::str_count(t, '\\*') - stringr::str_count(t, '\\-')) < opt$predictPCRartifacts_minReportHalfMatches){
+            p <- NA
+          } else if (stringr::str_count(p, '\\*') < opt$predictPCRartifacts_minReportMatches) {
+            p <- NA
+          } else {
+            p <- paste0(stringr::str_count(p, '\\*'), ' | ', p)
+          }
         
-        p <- paste0(unlist(lapply(1:(opt$predictPCRartifacts_adjacentSeqLength+dashes), function(n){
-          n <- as.character(n)
-          if(n %in% names(p)) return(p[names(p) == n])
-          return('.')
-        })), collapse = '')
-        
-        # Look at the right half of the alignment and report it if there are >= opt$predictPCRartifacts_minReportHalfMatches matches.
-        t <- substr(p, ceiling(opt$predictPCRartifacts_adjacentSeqLength / 2)+1, nchar(p))
-        
-        if((stringr::str_count(t, '\\*') - stringr::str_count(t, '\\-')) < opt$predictPCRartifacts_minReportHalfMatches){
-          p <- NA
-        } else if (stringr::str_count(p, '\\*') < opt$predictPCRartifacts_minReportMatches) {
-          p <- NA
+          invisible(file.remove(list.files(file.path(opt$outputDir, opt$predictPCRartifacts_outputDir, 'dbs'), pattern = f, full.names = TRUE)))
+          return(tibble(uniqueSite = x$uniqueSite, PCRartifact2 = p))
         } else {
-          p <- paste0(stringr::str_count(p, '\\*'), ' | ', p)
+          invisible(file.remove(list.files(file.path(opt$outputDir, opt$predictPCRartifacts_outputDir, 'dbs'), pattern = f, full.names = TRUE)))
+          return(tibble(uniqueSite = x$uniqueSite, PCRartifact2 = NA))
         }
-        
-        invisible(file.remove(list.files(file.path(opt$outputDir, opt$predictPCRartifacts_outputDir, 'dbs'), pattern = f, full.names = TRUE)))
-        
-        return(tibble(uniqueSite = x$uniqueSite, PCRartifact2 = p))
       } else {
         invisible(file.remove(list.files(file.path(opt$outputDir, opt$predictPCRartifacts_outputDir, 'dbs'), pattern = f, full.names = TRUE)))
         return(tibble(uniqueSite = x$uniqueSite, PCRartifact2 = NA))
